[DEV-726] moved mail class to base directory
