+++ src/notification/entities/mail/user-mail.ts
@@ -1,6 +1,6 @@
 import { UserData } from 'src/user/models/user-data/user-data.entity';
 import { NotificationMetadata, NotificationOptions } from '../notification.entity';
-import { Mail } from './mail';
+import { Mail } from './base/mail';
 
 export interface UserMailInput {
   userData: UserData;
